Move stuff into stan.requests.register subpackage
