+++ src/main/java/uk/ac/sanger/sccp/stan/service/register/RegisterService.java
@@ -1,8 +1,8 @@
 package uk.ac.sanger.sccp.stan.service.register;
 
 import uk.ac.sanger.sccp.stan.model.User;
-import uk.ac.sanger.sccp.stan.request.RegisterRequest;
-import uk.ac.sanger.sccp.stan.request.RegisterResult;
+import uk.ac.sanger.sccp.stan.request.register.RegisterRequest;
+import uk.ac.sanger.sccp.stan.request.register.RegisterResult;
 
 public interface RegisterService {
     RegisterResult register(RegisterRequest request, User user);
