+++ src/test/java/uk/ac/sanger/sccp/stan/service/register/TestRegisterValidation.java
@@ -11,8 +11,8 @@ import org.mockito.verification.VerificationMode;
 import uk.ac.sanger.sccp.stan.EntityFactory;
 import uk.ac.sanger.sccp.stan.model.*;
 import uk.ac.sanger.sccp.stan.repo.*;
-import uk.ac.sanger.sccp.stan.request.BlockRegisterRequest;
-import uk.ac.sanger.sccp.stan.request.RegisterRequest;
+import uk.ac.sanger.sccp.stan.request.register.BlockRegisterRequest;
+import uk.ac.sanger.sccp.stan.request.register.RegisterRequest;
 import uk.ac.sanger.sccp.stan.service.Validator;
 import uk.ac.sanger.sccp.stan.service.register.RegisterValidationImp.StringIntKey;
